+++ utils/fast_rrt/src/fast_rrt.cpp
@@ -3,25 +3,25 @@
 FastRRT::FastRRT(
     int og_width,
     int og_height,
-    float og_real_width_m,
-    float og_real_height_m,
+    double og_real_width_m,
+    double og_real_height_m,
     int min_dist_x,
     int min_dist_z,
     int lower_bound_x,
     int lower_bound_z,
     int upper_bound_x,
     int upper_bound_z,
-    float max_steering_angle,
-    float velocity_m_s)
+    double max_steering_angle,
+    double velocity_m_s)
 {
-    float rw = og_width / og_real_width_m;
-    float rh = og_height / og_real_height_m;
+    double rw = og_width / og_real_width_m;
+    double rh = og_height / og_real_height_m;
 
-    float3 center;
+    double3 center;
     center.x = static_cast<int>(round(og_width / 2));
     center.y = static_cast<int>(round(og_height / 2));
 
-    float lr = 0.5 * (lower_bound_z - upper_bound_z) / (og_height / og_real_height_m);
+    double lr = 0.5 * (lower_bound_z - upper_bound_z) / (og_height / og_real_height_m);
 
     _curveGenerator = new CurveGenerator(
         center,
@@ -51,15 +51,15 @@ FastRRT::~FastRRT()
     delete _curveGenerator;
 }
 
-Memlist<float3> *FastRRT::buildCurveWaypoints(float3 start, float velocity_meters_per_s, float steering_angle_deg, float path_size)
+Memlist<double3> *FastRRT::buildCurveWaypoints(double3 start, double velocity_meters_per_s, double steering_angle_deg, double path_size)
 {
     return _curveGenerator->buildCurveWaypoints(start, velocity_meters_per_s, steering_angle_deg, path_size);
 }
-Memlist<float3> *FastRRT::buildCurveWaypoints(float3 start, float3 end, float velocity_meters_per_s)
+Memlist<double3> *FastRRT::buildCurveWaypoints(double3 start, double3 end, double velocity_meters_per_s)
 {
     return _curveGenerator->buildCurveWaypoints(start, end, velocity_meters_per_s);
 }
 
-void FastRRT::testDrawPath(float3 *og, float3 &start, float3 &end) {
+void FastRRT::testDrawPath(float3 *og, double3 &start, double3 &end) {
     _graph->drawKinematicPath(og, start, end);
 }
